\name{vlist}
\alias{vlist}
\docType{data}
\title{ Sample experimental/phenotype data for LMGene package}
\description{
  List of experimental factors for the sample matrix array data, 'sample.mat'.
}
\usage{data(vlist)}
\examples{
library(Biobase)
library(LMGene)

#data
data(vlist)

vlist

}
\keyword{datasets}
