\name{vlist}
\alias{vlist}
\docType{data}
\title{ Sample experimental data for LMGene package}
\description{
  A list data representing experiment description information for the sample matrix array data, 'sample.mat'.
}
\usage{data(vlist)}
\examples{
#library
library(Biobase)
library(LMGene)

#data
data(vlist)

vlist

}
\keyword{datasets}
